+++ packages/@coorpacademy-progression-engine/src/common/test/check-answer-correctness.qcm-graphic.js
