+++ packages/@coorpacademy-progression-engine/src/common/test/fixtures/slides.js
