+++ packages/@coorpacademy-progression-engine/src/engines/learner/test/config.js
@@ -1,8 +1,8 @@
 import test from 'ava';
 import uniqBy from 'lodash/fp/uniqBy';
-import microlearning from '../microlearning';
+import config from '../config';
 
-const {configurations} = microlearning;
+const {configurations} = config;
 
 test("Configurations shouldn't have conflicting version numbers", t => {
   t.is(configurations.length, uniqBy('version', configurations).length);
