+++ packages/@coorpacademy-progression-engine/src/engines/microlearning/compute-next-step.js
@@ -1,0 +1,3 @@
+import computeNextStep from '../learner/compute-next-step';
+
+export default computeNextStep;
