+++ packages/@coorpacademy-progression-engine/src/engines/microlearning/config.js
@@ -1,11 +1,10 @@
 // @flow
-import type {Config} from '../types';
+import type {Config} from './types';
 
 const realConfigurations: Array<Config> = [
   {
     version: '1',
     lives: 1,
-    livesDisabled: false,
     maxTypos: 2,
     slidesToComplete: 4,
     answerBoundaryLimit: 5,
@@ -20,7 +19,6 @@ const testConfigurations: Array<Config> = [
   {
     version: 'allow_typos_3',
     lives: 1,
-    livesDisabled: false,
     maxTypos: 3,
     slidesToComplete: 4,
     answerBoundaryLimit: 5,
