+++ packages/@coorpacademy-progression-engine/src/engines/microlearning/test/config.js
@@ -1,8 +1,8 @@
 import test from 'ava';
 import uniqBy from 'lodash/fp/uniqBy';
-import learner from '../learner';
+import config from '../config';
 
-const {configurations} = learner;
+const {configurations} = config;
 
 test("Configurations shouldn't have conflicting version numbers", t => {
   t.is(configurations.length, uniqBy('version', configurations).length);
