+++ packages/@coorpacademy-progression-engine/src/engines/microlearning/update-state.js
@@ -1,0 +1,3 @@
+import updateState from '../learner/update-state';
+
+export default updateState;
