+++ packages/@coorpacademy-progression-engine/src/index.js
@@ -1,12 +1,14 @@
 // @flow
-import computeNextStep from './compute-next-step';
-import checkAnswer from './check-answer';
-import checkAnswerCorrectness from './check-answer-correctness';
-import updateState from './update-state';
-import createProgression from './create-progression';
-import getConfig from './config';
+import combineReducers from './common/combine-reducers';
+import computeNextStep from './engines/learner/compute-next-step';
+import checkAnswer from './common/check-answer';
+import checkAnswerCorrectness from './common/check-answer-correctness';
+import updateState from './engines/learner/update-state';
+import createProgression from './common/create-progression';
+import getConfig from './common/config';
 
 export {
+  combineReducers,
   computeNextStep,
   checkAnswer,
   checkAnswerCorrectness,
